Move `TickFuture` to its own module
